refactor(parser): split stmt/sub.rs into cohesive submodules (§7-8)
Split the 2347-line parser/stmt/sub.rs (over the 500-line limit) into a
facade plus seven themed submodules under src/parser/stmt/sub/, all under
500 lines. sub.rs is entirely free parser functions addressed as
sub::<name> (25 call sites); the facade re-exports each submodule
preserving its original visibility, so no external caller changes.
stmt/mod.rs is untouched.

- sub.rs (62)             - facade: use lines + re-exports + sub_param forwarders
- sub_decl.rs (452)       - sub/anon-multi/main-semicolon decls + body
- param_validate.rs (385) - param-trait + signature validation
- traits.rs (384)         - parse_sub_traits + export/attr trait checks
- param_list.rs (302)     - parse_param_list + duplicate/sigil checks
- sub_name.rs (300)       - sub-name resolution
- return_type.rs (294)    - --> return-type parsing
- op_name.rs (222)        - operator-name (:[...]/quoted) helpers

Pure relocation. Moved fns are pub(crate) in submodules and re-exported at
their original visibility; seven cross-submodule private helpers bumped to
pub(crate). Body references to stmt-level siblings gained one super:: level.
The fn multiset is byte-identical. No behavior change.
